requirements.txt
word2vec
numpy
scipy==1.1.0
pillow
lmdb
torchfile
scikit-learn
tqdm
nltk
simplejson
torch==1.6.0
torchvision==0.7.0
pillow
matplotlib
tensorboard
tensorboardX
lmdb
tfrecord